pyproject.toml
[tool.poetry]
name = "telegram-drive-bot"
version = "0.1.0"
description = ""
authors = ["oCa"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "21.6"
google-auth = "2.35.0"
google-auth-oauthlib = "1.2.1"
google-api-python-client = "2.147.0"
oauth2client = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
